refactor and split out into multiple modules
